requirements.txt
websockets>=12.0
pyyaml>=6.0
aiofiles>=23.0
aiohttp>=3.9.0
